Use symlinks
